Added JPEG and oversize image support
FPBioimage will first look for a PNG image with the name specified in
the javascript parameters. If it can’t find a PNG image, it will look
for a .jpg file with the same name.
Images larger than 500x500x500 are now down-sampled in FPBioimage -
although this should only be used for advanced applications, to save on
the amount of data which has to be sent to the user.